Creates PointController and moves render logic of the card there
